Address GPT-5.5 review: fix lifetime, overflow, marker and table bugs
- #1 Lifetime: the C-API backend now bases each exported NumPy array on a
  per-read capsule that owns the backing jstruct, so arrays stay valid after
  the object is dropped or re-read into. The Cython backend bases its arrays
  on the DecompressedJpeg (via the NumPy C-API) instead of leaving them
  ownerless.
- #2 mat2D allocates and indexes with size_t, preventing signed-int overflow
  (and the resulting under-allocation) on maliciously large JPEG dimensions.
- #3/#6 Markers are stored as std::vector<std::string> with their exact
  length: no more invalid delete[] of borrowed pointers, and binary COM
  markers with embedded NUL bytes now round-trip correctly.
- #4 Original quant/Huffman table slot numbers (0..3) are recorded on read and
  restored on write, so JPEGs with sparse table slots round-trip.
- #5 jpeg_write now populates freshly allocated Huffman tables.
- #7 get_coef_block_array_shape validates the channel index.
- #8 read_huff_list checks PyDict_SetItemString failures.
- #9 setup.py raises a clear error when JPEGIO_BACKEND=cython is built without
  Cython; README documents the requirement.

Verified: both backends build and pass the full test suite on Python 3.9 and
3.13, plus new re-read / del-after / binary-marker / bounds checks.

+++ src/jpegio/_backend/mat2D.h
@@ -30,14 +30,16 @@ public:
 	{
 		this->rows = rows;
 		this->cols = cols;
-		this->vect = std::vector<T>(rows*cols);
+		// size_t multiply avoids signed-int overflow for large (e.g. malicious)
+		// JPEG dimensions, which would otherwise under-allocate the buffer.
+		this->vect = std::vector<T>((size_t)rows * (size_t)cols);
 	}
 
 	mat2D(int rows, int cols, T baseVal)
 	{
 		this->rows = rows;
 		this->cols = cols;
-		this->vect = std::vector<T>(rows*cols, baseVal);
+		this->vect = std::vector<T>((size_t)rows * (size_t)cols, baseVal);
 	}
 
 	~mat2D()
@@ -47,7 +49,7 @@ public:
 
 	T Read(int row, int col)
 	{
-		return this->vect[row*cols+col];
+		return this->vect[(size_t)row * cols + col];
 	}
 
 	int valueCount(T val)
@@ -57,7 +59,7 @@ public:
 
 	void Write(int row, int col, T val)
 	{
-		this->vect[row*cols+col] = val;
+		this->vect[(size_t)row * cols + col] = val;
 	}
 
 	void PermuteElements()
